+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -8,6 +8,8 @@ import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.exception.DuplicatedStationNameException;
+import wooteco.subway.station.exception.NoSuchStationException;
 import wooteco.subway.station.exception.StationAlreadyRegisteredInLineException;
 
 import java.util.List;
@@ -21,11 +23,17 @@ public class StationService {
 
     @Transactional
     public StationResponse saveStation(StationRequest stationRequest) {
+        if (stationDao.isStationExist(stationRequest.getName())) {
+            throw new DuplicatedStationNameException();
+        }
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
 
     public Station findStationById(Long id) {
+        if (stationDao.isStationNotExist(id)) {
+            throw new NoSuchStationException();
+        }
         return stationDao.findById(id);
     }
 
@@ -42,6 +50,9 @@ public class StationService {
     public void deleteStationById(Long id) {
         if (sectionDao.isStationExistInSection(id)) {
             throw new StationAlreadyRegisteredInLineException();
+        }
+        if (stationDao.isStationNotExist(id)) {
+            throw new NoSuchStationException();
         }
         stationDao.deleteById(id);
     }
